test(e2e): keep tunnel lifecycle test linear
